backend/requirements.txt
sqlalchemy==1.2.10
flask-login==0.3.2
Flask-Script==2.0.6
pymysql==0.9.3